+++ src/Blitz.php
@@ -8,7 +8,6 @@ namespace putyourlightson\blitz;
 use Craft;
 use craft\base\Element;
 use craft\base\Plugin;
-use craft\console\controllers\ResaveController;
 use craft\elements\db\ElementQuery;
 use craft\elements\User;
 use craft\enums\CmsEdition;
@@ -24,6 +23,8 @@ use craft\events\RegisterUrlRulesEvent;
 use craft\events\RegisterUserPermissionsEvent;
 use craft\helpers\UrlHelper;
 use craft\log\MonologTarget;
+use craft\queue\jobs\ResaveElements;
+use craft\queue\Queue as CraftQueue;
 use craft\services\Dashboard;
 use craft\services\Elements;
 use craft\services\Plugins;
@@ -60,11 +61,11 @@ use putyourlightson\blitz\utilities\HintsUtility;
 use putyourlightson\blitz\variables\BlitzVariable;
 use putyourlightson\blitz\widgets\CacheWidget;
 use putyourlightson\sprig\Sprig;
-use yii\base\Controller;
 use yii\base\Event;
 use yii\di\Instance;
 use yii\log\Dispatcher;
 use yii\log\Logger;
+use yii\queue\ExecEvent;
 use yii\queue\Queue;
 
 /**
@@ -412,9 +413,7 @@ class Blitz extends Plugin
         $events = [
             [Elements::class, Elements::EVENT_BEFORE_RESAVE_ELEMENTS],
             [Elements::class, Elements::EVENT_BEFORE_PROPAGATE_ELEMENTS],
-            [ResaveController::class, Controller::EVENT_BEFORE_ACTION],
         ];
-
         foreach ($events as $event) {
             Event::on($event[0], $event[1],
                 function() {
@@ -423,13 +422,19 @@ class Blitz extends Plugin
             );
         }
 
+        Event::on(Queue::class, Queue::EVENT_BEFORE_EXEC,
+            function(ExecEvent $event) {
+                if ($event->job instanceof ResaveElements) {
+                    $this->refreshCache->batchMode = true;
+                }
+            }
+        );
+
         // Refresh the cache
         $events = [
             [Elements::class, Elements::EVENT_AFTER_RESAVE_ELEMENTS],
             [Elements::class, Elements::EVENT_AFTER_PROPAGATE_ELEMENTS],
-            [ResaveController::class, Controller::EVENT_AFTER_ACTION],
         ];
-
         foreach ($events as $event) {
             Event::on($event[0], $event[1],
                 function() {
@@ -437,6 +442,14 @@ class Blitz extends Plugin
                 }
             );
         }
+
+        Event::on(CraftQueue::class, CraftQueue::EVENT_AFTER_EXEC_AND_RELEASE,
+            function(ExecEvent $event) {
+                if ($event->job instanceof ResaveElements) {
+                    $this->refreshCache->refresh();
+                }
+            }
+        );
     }
 
     /**